integration-examples/pipedrive-lookalike-companies/requirements.txt
python-dotenv>=1.0.0
-r ../../core/surfe/requirements.txt
-r ../../core/integrations/pipedrive/requirements.txt 